requirements.txt
APScheduler==3.6.3
black==23.3.0
cachetools==4.2.2
certifi==2023.5.7
click==8.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
psycopg2-binary==2.9.6
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2023.3
six==1.16.0
SQLAlchemy==1.4.39
tornado==6.3.2
tzlocal==5.0.1
